The accessors/modifiers are now done through XS.
Before the object wasn't registered as a GType and all marshalling had to be
done by hand. Furthermore the C structure was transformed into a Per hash
making the hash fields out of sync with the C structure. Now that the object is
a GType (GBoxed) everything is more natural.