Fix 591132: Attempting to remove actor when zooming in/out
This warning happened when the cache was being validated and
that a newer version of the tile was available.  Plus,
the ChamplainTile had to be in the process of being destroyed
(happens if you zoom 2 levels down quickly).  The new version
of the tile was then fading in on a non existing ClutterActor.

I added checks so that
* the animation is not created in such a case,
* the animation is forced to complete on dispose of ChamplainTiles
* unrelated but: only one animation can happen on a tile at one,
  the previously running animation will be forced to complete before
  the new one is setup.