Fix 590727: Events are not send until the mouse move
Do not grab the mouse when the source actor is not the current actor,
resulting in the event being passed down, but also in that you cannot
drag the map starting on an actor anymore).

Connecting the signal handlers on an object at the same or lower level
than fingerscroll fixed the transmission of events.