Make it possible to manually chain chains.
Chains should behave as other map sources so it should be possible
to create chains of chains. For now, chains of chains have to be
created manually - i.e. the outer chain cannot be ChamplainMapSourceChain
but has to be created by _set_next_source(). The reason is that
a chain has to know about each of its elements whether it is a map
source or a cache (if this feature is desirable, we can add a virtual
method of map sources that determines whether it behaves as a map source
or cache so even if the map source is a chain, it could have this
information attached).

Signed-off-by: Jiří Techet <techet@gmail.com>