view: Remove the need to explicitly set the size on the actor
Firstly we need to report a size for the map using the ClutterActor
_get_preferred_width/_height virtual functions. For both the minimum and
natural dimensions we use the tile size which is the smallest sensible size
for the map.

We also need to save the dimensions of the allocation box for the actor and
save that into the private structure. We then used that in an idle to update
the content to show.