Insert tile content into parent only when state set to DONE
We call set_content() on tile also when we want to validate the tile only
and if the tile is not validated, the new contents is set again by
calling set_content(). This would cause redrawing the tile twice instead
of once. Also the tile is repositioned only after the state is set to
DONE so in between the tile is displayed at incorrect place.

Signed-off-by: Jiří Techet <techet@gmail.com>