Base viewport update on time as well
Viewport is currently updated only based on the amount of scrolled pixels
(100). However, when the viewport is slowing down in the kinetic mode,
this can result in a blank space appearing in the top of the window for
quite a long time because the movement is slow and doesn't reach the 100
pixel margin needed for update. This patch adds one more criterion - time,
which makes to perform the updates at least 4 times a second, no matter
how many pixels were scrolled.

Signed-off-by: Jiří Techet <techet@gmail.com>