Move fill_tile() call from view_load_visible_tiles() to an idle callback
Even though we load tiles asynchronously, the preparation steps take
some time. Inside the inner loop of view_load_visible_tiles() the time
is multiplied by the number of loaded tiles. For improved interactivity it
is better to load tiles outside of the loop in idle functions so it can be
interleaved with other main loop calls. This significantly improves
interactivity in full screen launcher-gtk demo.

Signed-off-by: Jiří Techet <techet@gmail.com>

add commit