Optimize ChamplainView state update
When profiling with Oprofile and gprof2dot, view_update_state() consumed
nearly one half of CPU cycles (CPU used by libchamplain only without
any other libraries). This was because we were cycling through all
the tiles to determine the state and the state was updated for every
tile, which lead to quadratic complexity. This patch fixes this by
updating a counter of tiles that are being filled.

Signed-off-by: Jiří Techet <techet@gmail.com>