Layer speedup
The following patch replaces marker sorting ( by latitude ) with
clutter's set_depth to achieve the same result.

Using a modified animated-marker demo with 484 markers, this patch
reduces CPU using on my laptop from 80% to 19%.

There is still more that can be done, but that's now in clutter, which
appears to be sorting a list every time a depth is changed.