Do not order markers according to latitude
The marker ordering using the z coordinate causes that markers are not
located exactly above their location - this is caused by the perspective
projection. This also causes that the markers displayed in the front
are bigger than the markers displayed below them. Because the previous
ordering implementation was too slow and this implementation causes serious
issues, this commit removes the marker ordering completely.

Signed-off-by: Jiří Techet <techet@gmail.com>