Fire latitude and longitude signals independently of viewport shift
The check for distance of latitude and longitude change is not necessary
because the function update_viewport() is already called from
viewport_pos_changed_cb() only 4-times a second (or when the distance is
larger than 100).

This patch also replaces the use of g_timeout() by g_timer_elapsed(), which
makes the code nicer (thanks to Tomaž Vajngerl).