Merge ChamplainLayer and ChamplainPolygon
In principle, ChamplainLayer and ChamplainPolygon do the same thing.
They both contain a list of points on the map (ChamplainMarker or
ChamplainPoint) which in the case of ChamplainPolygon are defining the
polygon. By merging both concept, we can reuse a lot of code, such as
marker/polygon node moving, all point manipulation and so on.