Perform fade-in animation on ChamplainTile and not the containing actor
Under GTK 3 if we perform the animation on the content actor,
the animation doesn't work properly. For now it remains mystery
why this happens. At the moment we are fine with this solution
but if we implement some better zoom-in/out animation that
fades-in the loaded tile into the scaled tile from the previous
zoom level, we may need to animate the new content actor on top
of the old content actor inside the tile.