Finally fix our redrawing problem
This one fixes our hard-to-find redrawing problem. What happened:
in the paint method of the viewport the coordinates were transformed
only for painting but clutter clips the images based on real coordinates
of the actors. Therefore clutter eliminated all the actors because it
thought they weren't displayed.

Instead of using low level cogl operations, use standard clutter actor
positioning to move the viewport - this fixes the problem completely.