ChamplainPoint: Port code to Clutter 1.12 API
There is still a call to deprecated clutter_actor_set_anchor_point() but
I don't yet know how to replace that one. According to docs it should
simply be replaced by clutter_actor_set_pivot_point but seem translation
isn't that straightforward as anchor and pivot point are not the same.

https://bugzilla.gnome.org/show_bug.cgi?id=697711