view: Implement and add horizontal-wrap property
To give the impression of horizontal wrap clones of the map- and
user layers will be placed next to the real layers. When we cross over
to them we will re-position the viewport to give the illusion of wrapping
the map. The same trick is needed for the zoom actor.

https://bugzilla.gnome.org/show_bug.cgi?id=577597