Remove clones of user_layers
Clones of user_layers are totally passive and cannot be clicked or manipulated.
This will be very confusing to users as just one of the user layers
is master and can be clicked/moved/dragged and it's not clear which
one it is.

This patch removes the clones. TODO: move the master user_layers into the
center of the map.