view: Stop emission of ClutterZoomAction::gesture-cancelled
We successfully override ::gesture-begin and ::zoom in order to
replace the default "zoom the actor" behavior of ClutterZoomAction,
however ClutterZoomAction::gesture-cancel implementation has the side
effect of attempting to restore the actor translation and scale to the
original position, which is undesirable here.

This fixes the ChamplainView going blank, as it tries to "restore" a
0 scale, because we didn't give it an opportunity to fetch the actor
one on ClutterZoomAction::gesture-begin.

https://bugzilla.gnome.org/show_bug.cgi?id=757149