Don't think that left+right mouse buttons pressed mean multitouch
At the moment when both left and right mouse buttons are pressed, they
are regarded as multitouch and start of gesture-based zoom which leads
to strange zooming in/out while holding both mouse buttons and moving the
cursor.

Don't start gesture event when using mouse. (Maybe some of the other
devices might make sense here too but I don't have them to test.
Anyway, mouse is the most common one so this should be OK.)