meson: Fix gtk-doc module name
gtk-docs get installed into API versioned subdir (for being able to
parallel-install with past and potential future API versions), but
unversioned package_name is passed to gnome.gtkdoc. This results in
/usr/share/gtk-doc/html/champlain-0.12/champlain.devhelp2 being
installed, which the devhelp tool ignores. The devhelp file needs to
be named the same as the subdir for it to be picked up, thus use
package_string instead of package_name as gtkdoc positional argument
to fix it.