Make the build reproducible
Whilst working on the Reproducible Builds effort [0] we noticed
that libchamplain could not be built reproducibly. This is because
it used the absolute build path in the comments of the .h files
generated my glib-mkenums.

Patch attached. It also changes the equivalent .c file for
completeness but this is not strictly required.

 [0] https://reproducible-builds.org/

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>